+++ packages/cactus-plugin-ledger-connector-fabric/src/main/typescript/plugin-ledger-connector-fabric.ts
@@ -3,6 +3,7 @@ import path from "path";
 import { Server } from "http";
 import { Server as SecureServer } from "https";
 
+import { Certificate } from "@fidm/x509";
 import { Express } from "express";
 import "multer";
 import temp from "temp";
@@ -17,10 +18,10 @@ import {
   DefaultEventHandlerStrategies,
   Gateway,
   GatewayOptions,
-  Identity,
-  InMemoryWallet,
-  X509WalletMixin,
+  Wallets,
+  X509Identity,
   TransientMap,
+  Wallet,
 } from "fabric-network";
 
 import { Optional } from "typescript-optional";
@@ -69,6 +70,8 @@ import {
   RunTransactionResponse,
   ChainCodeProgrammingLanguage,
   ChainCodeLifeCycleCommandResponses,
+  FabricSigningCredential,
+  DefaultEventHandlerStrategy,
 } from "./generated/openapi/typescript-axios/index";
 
 import {
@@ -85,6 +88,8 @@ import {
 } from "./deploy-contract/deploy-contract-endpoint-v1";
 import { sourceLangToRuntimeLang } from "./peer/source-lang-to-runtime-lang";
 import FabricCAServices from "fabric-ca-client";
+import { createGateway } from "./common/create-gateway";
+import { Endorser } from "fabric-common";
 
 /**
  * Constant value holding the default $GOPATH in the Fabric CLI container as
@@ -802,14 +807,92 @@ export class PluginLedgerConnectorFabric
     return endpoints;
   }
 
+  protected async createGateway(req: RunTransactionRequest): Promise<Gateway> {
+    if (req.gatewayOptions) {
+      return createGateway({
+        logLevel: this.opts.logLevel,
+        pluginRegistry: this.opts.pluginRegistry,
+        defaultConnectionProfile: this.opts.connectionProfile,
+        defaultDiscoveryOptions: this.opts.discoveryOptions || {
+          enabled: true,
+          asLocalhost: true,
+        },
+        defaultEventHandlerOptions: this.opts.eventHandlerOptions || {
+          endorseTimeout: 300,
+          commitTimeout: 300,
+          strategy: DefaultEventHandlerStrategy.NetworkScopeAllfortx,
+        },
+        gatewayOptions: req.gatewayOptions,
+      });
+    } else {
+      return this.createGatewayLegacy(req.signingCredential);
+    }
+  }
+
+  protected async createGatewayLegacy(
+    signingCredential: FabricSigningCredential,
+  ): Promise<Gateway> {
+    const { connectionProfile, eventHandlerOptions: eho } = this.opts;
+
+    const wallet = await Wallets.newInMemoryWallet();
+
+    const keychain = this.opts.pluginRegistry.findOneByKeychainId(
+      signingCredential.keychainId,
+    );
+    this.log.debug(
+      "transact() obtained keychain by ID=%o OK",
+      signingCredential.keychainId,
+    );
+
+    const fabricX509IdentityJson = await keychain.get(
+      signingCredential.keychainRef,
+    );
+    this.log.debug(
+      "transact() obtained keychain entry Key=%o OK",
+      signingCredential.keychainRef,
+    );
+    const identity = JSON.parse(fabricX509IdentityJson);
+
+    await wallet.put(signingCredential.keychainRef, identity);
+    this.log.debug("transact() imported identity to in-memory wallet OK");
+
+    const eventHandlerOptions: DefaultEventHandlerOptions = {
+      commitTimeout: this.opts.eventHandlerOptions?.commitTimeout || 300,
+      endorseTimeout: 300,
+    };
+    if (eho?.strategy) {
+      eventHandlerOptions.strategy =
+        DefaultEventHandlerStrategies[eho.strategy];
+    }
+
+    const gatewayOptions: GatewayOptions = {
+      discovery: this.opts.discoveryOptions,
+      eventHandlerOptions,
+      identity: signingCredential.keychainRef,
+      wallet,
+    };
+
+    this.log.debug(`discovery=%o`, gatewayOptions.discovery);
+    this.log.debug(`eventHandlerOptions=%o`, eventHandlerOptions);
+
+    const gateway = new Gateway();
+
+    await gateway.connect(
+      connectionProfile as ConnectionProfile,
+      gatewayOptions,
+    );
+
+    this.log.debug("transact() gateway connection established OK");
+
+    return gateway;
+  }
+
   public async transact(
     req: RunTransactionRequest,
   ): Promise<RunTransactionResponse> {
     const fnTag = `${this.className}#transact()`;
 
-    const { connectionProfile, eventHandlerOptions: eho } = this.opts;
     const {
-      signingCredential,
       channelName,
       contractName,
       invocationType,
@@ -819,53 +902,11 @@ export class PluginLedgerConnectorFabric
       endorsingParties,
     } = req;
 
-    const gateway = new Gateway();
-    const wallet = new InMemoryWallet(new X509WalletMixin());
-    const keychain = this.opts.pluginRegistry.findOneByKeychainId(
-      signingCredential.keychainId,
-    );
-    this.log.debug(
-      "transact() obtained keychain by ID=%o OK",
-      signingCredential.keychainId,
-    );
-
-    const fabricX509IdentityJson = await keychain.get<string>(
-      signingCredential.keychainRef,
-    );
-    this.log.debug(
-      "transact() obtained keychain entry Key=%o OK",
-      signingCredential.keychainRef,
-    );
-    const identity = JSON.parse(fabricX509IdentityJson);
-
     try {
-      await wallet.import(signingCredential.keychainRef, identity);
-      this.log.debug("transact() imported identity to in-memory wallet OK");
-
-      const eventHandlerOptions: DefaultEventHandlerOptions = {
-        commitTimeout: this.opts.eventHandlerOptions?.commitTimeout || 300,
-      };
-      if (eho?.strategy) {
-        eventHandlerOptions.strategy =
-          DefaultEventHandlerStrategies[eho.strategy];
-      }
-
-      const gatewayOptions: GatewayOptions = {
-        discovery: this.opts.discoveryOptions,
-        eventHandlerOptions,
-        identity: signingCredential.keychainRef,
-        wallet,
-      };
-
-      this.log.debug(`discovery=%o`, gatewayOptions.discovery);
-      this.log.debug(`eventHandlerOptions=%o`, eventHandlerOptions);
-      await gateway.connect(
-        connectionProfile as ConnectionProfile,
-        gatewayOptions,
-      );
-      this.log.debug("transact() gateway connection established OK");
-
+      const gateway = await this.createGateway(req);
       const network = await gateway.getNetwork(channelName);
+      // const channel = network.getChannel();
+      // const endorsers = channel.getEndorsers();
       const contract = network.getContract(contractName);
 
       let out: Buffer;
@@ -877,7 +918,37 @@ export class PluginLedgerConnectorFabric
           break;
         }
         case FabricContractInvocationType.Send: {
-          out = await contract.submitTransaction(fnName, ...params);
+          const tx = contract.createTransaction(fnName);
+          if (req.endorsingPeers) {
+            const { endorsingPeers } = req;
+            const channel = network.getChannel();
+
+            const allChannelEndorsers = (channel.getEndorsers() as unknown) as Array<
+              Endorser & { options: { pem: string } }
+            >;
+
+            const endorsers = allChannelEndorsers
+              .map((endorser) => {
+                const certificate = Certificate.fromPEM(
+                  (endorser.options.pem as unknown) as Buffer,
+                );
+                return { certificate, endorser };
+              })
+              .filter(
+                ({ endorser, certificate }) =>
+                  endorsingPeers.includes(endorser.mspid) ||
+                  endorsingPeers.includes(certificate.issuer.organizationName),
+              )
+              .map((it) => it.endorser);
+
+            this.log.debug(
+              "%o endorsers: %o",
+              endorsers.length,
+              endorsers.map((it) => `${it.mspid}:${it.name}`),
+            );
+            tx.setEndorsingPeers(endorsers);
+          }
+          out = await tx.submit(...params);
           success = true;
           break;
         }
@@ -967,23 +1038,30 @@ export class PluginLedgerConnectorFabric
     mspId: string,
     enrollmentID: string,
     enrollmentSecret: string,
-  ): Promise<[Identity, InMemoryWallet]> {
+  ): Promise<[X509Identity, Wallet]> {
     const fnTag = `${this.className}#enrollAdmin()`;
     try {
       const ca = await this.createCaClient(caId);
-      const wallet = new InMemoryWallet(new X509WalletMixin());
+      const wallet = await Wallets.newInMemoryWallet();
 
       // Enroll the admin user, and import the new identity into the wallet.
       const request = { enrollmentID, enrollmentSecret };
       const enrollment = await ca.enroll(request);
 
-      const { certificate: cert, key } = enrollment;
+      const { certificate, key } = enrollment;
       const keyBytes = key.toBytes();
 
-      const identity = X509WalletMixin.createIdentity(mspId, cert, keyBytes);
-      await wallet.import(identityId, identity);
+      const x509Identity: X509Identity = {
+        credentials: {
+          certificate,
+          privateKey: keyBytes,
+        },
+        mspId,
+        type: "X.509",
+      };
+      await wallet.put(identityId, x509Identity);
 
-      return [identity, wallet];
+      return [x509Identity, wallet];
     } catch (ex) {
       this.log.error(`enrollAdmin() Failure:`, ex);
       throw new Error(`${fnTag} Exception: ${ex?.message}`);