+++ packages/cactus-test-plugin-htlc-eth-besu-erc20/src/test/typescript/integration/plugin-htlc-eth-besu-erc20/get-single-status-endpoint-invalid.test.ts
@@ -38,7 +38,8 @@ import { PluginKeychainMemory } from "@hyperledger/cactus-plugin-keychain-memory
 const logLevel: LogLevelDesc = "INFO";
 const estimatedGas = 6721975;
 const expiration = 2147483648;
-const receiver = "0x627306090abaB3A6e1400e9345bC60c78a8BEf57";
+const besuTestLedger = new BesuTestLedger({ logLevel });
+const receiver = "0x" + besuTestLedger.getGenesisAccountPubKey();
 const hashLock =
   "0x3c335ba7f06a8b01d0596589f73c19069e21c81e5013b91f408165d1bf623d32";
 const firstHighNetWorthAccount = "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1";
@@ -63,9 +64,10 @@ test(testCase, async (t: Test) => {
   t.comment("Starting Besu Test Ledger");
   const besuTestLedger = new BesuTestLedger({ logLevel });
 
-  test.onFailure(async () => {
+  test.onFinish(async () => {
     await besuTestLedger.stop();
     await besuTestLedger.destroy();
+    await pruneDockerAllIfGithubAction({ logLevel });
   });
   await besuTestLedger.start();
 
@@ -78,11 +80,19 @@ test(testCase, async (t: Test) => {
     // pre-provision keychain with mock backend holding the private key of the
     // test account that we'll reference while sending requests with the
     // signing credential pointing to this keychain entry.
-    backend: new Map([[TestTokenJSON.contractName, TestTokenJSON]]),
+    backend: new Map([
+      [TestTokenJSON.contractName, JSON.stringify(TestTokenJSON)],
+    ]),
     logLevel,
   });
-  keychainPlugin.set(DemoHelperJSON.contractName, DemoHelperJSON);
-  keychainPlugin.set(HashTimeLockJSON.contractName, HashTimeLockJSON);
+  keychainPlugin.set(
+    DemoHelperJSON.contractName,
+    JSON.stringify(DemoHelperJSON),
+  );
+  keychainPlugin.set(
+    HashTimeLockJSON.contractName,
+    JSON.stringify(HashTimeLockJSON),
+  );
 
   const factory = new PluginFactoryLedgerConnector({
     pluginImportType: PluginImportType.Local,
@@ -223,12 +233,12 @@ test(testCase, async (t: Test) => {
     web3SigningCredential,
     gas: estimatedGas,
   };
-  const resNewContract = await api.newContract(request);
+  const resNewContract = await api.newContractV1(request);
   t.equal(resNewContract.status, 200, "response status is OK");
   t.comment("Get status with invalid id");
 
   const fakeId = "0x66616b654964";
-  const res = await api.getSingleStatus(
+  const res = await api.getSingleStatusV1(
     fakeId,
     web3SigningCredential,
     connectorId,
@@ -236,14 +246,5 @@ test(testCase, async (t: Test) => {
   );
   t.equal(res.status, 200, "response status is 200 OK");
   t.equal(res.data, 0, "the contract status is 0 - INVALID");
-  await besuTestLedger.stop();
-  await besuTestLedger.destroy();
-
-  t.end();
-});
-
-test("AFTER " + testCase, async (t: Test) => {
-  const pruning = pruneDockerAllIfGithubAction({ logLevel });
-  await t.doesNotReject(pruning, "Pruning did not throw OK");
   t.end();
 });