+++ packages/cactus-test-plugin-htlc-eth-besu-erc20/src/test/typescript/integration/plugin-htlc-eth-besu-erc20/initialize-endpoint.test.ts
@@ -62,6 +62,7 @@ test(testCase, async (t: Test) => {
     await besuTestLedger.stop();
     t.comment("Ledger #1 destroying...");
     await besuTestLedger.destroy();
+    await pruneDockerAllIfGithubAction({ logLevel });
   });
 
   const rpcApiHttpHost = await besuTestLedger.getRpcApiHttpHost();
@@ -73,7 +74,9 @@ test(testCase, async (t: Test) => {
     // pre-provision keychain with mock backend holding the private key of the
     // test account that we'll reference while sending requests with the
     // signing credential pointing to this keychain entry.
-    backend: new Map([[HashTimeLockJSON.contractName, HashTimeLockJSON]]),
+    backend: new Map([
+      [HashTimeLockJSON.contractName, JSON.stringify(HashTimeLockJSON)],
+    ]),
     logLevel,
   });
 
@@ -131,7 +134,7 @@ test(testCase, async (t: Test) => {
     gas: estimatedGas,
   };
 
-  const res = await api.initialize(request);
+  const res = await api.initializeV1(request);
   t.equal(res.status, 200, "response status is 200 OK");
 
   t.ok(res.data, "pluginHtlc.initialize() output is truthy OK");
@@ -168,7 +171,9 @@ test("Test initialize function with invalid params", async (t: Test) => {
     // pre-provision keychain with mock backend holding the private key of the
     // test account that we'll reference while sending requests with the
     // signing credential pointing to this keychain entry.
-    backend: new Map([[HashTimeLockJSON.contractName, HashTimeLockJSON]]),
+    backend: new Map([
+      [HashTimeLockJSON.contractName, JSON.stringify(HashTimeLockJSON)],
+    ]),
     logLevel,
   });
 
@@ -227,16 +232,10 @@ test("Test initialize function with invalid params", async (t: Test) => {
     gas: estimatedGas,
   };
   try {
-    const res = await api.initialize(request);
+    const res = await api.initializeV1(request);
     t.equal(res.status, 400, "response status is 400");
   } catch (error) {
     t.equal(error.response.status, 400, "response status is 400");
   }
-  t.end();
-});
-
-test("AFTER " + testCase, async (t: Test) => {
-  const pruning = pruneDockerAllIfGithubAction({ logLevel });
-  await t.doesNotReject(pruning, "Pruning did not throw OK");
   t.end();
 });