+++ packages/cactus-test-plugin-htlc-eth-besu-erc20/src/test/typescript/integration/plugin-htlc-eth-besu-erc20/refund-endpoint.test.ts
@@ -39,7 +39,8 @@ import HashTimeLockJSON from "../../../../../../cactus-plugin-htlc-eth-besu-erc2
 
 const logLevel: LogLevelDesc = "INFO";
 const estimatedGas = 6721975;
-const receiver = "0x627306090abaB3A6e1400e9345bC60c78a8BEf57";
+const besuTestLedger = new BesuTestLedger({ logLevel });
+const receiver = "0x" + besuTestLedger.getGenesisAccountPubKey();
 const hashLock =
   "0x3c335ba7f06a8b01d0596589f73c19069e21c81e5013b91f408165d1bf623d32";
 const firstHighNetWorthAccount = "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1";
@@ -75,6 +76,7 @@ test(testCase, async (t: Test) => {
   test.onFinish(async () => {
     await besuTestLedger.stop();
     await besuTestLedger.destroy();
+    await pruneDockerAllIfGithubAction({ logLevel });
   });
 
   const rpcApiHttpHost = await besuTestLedger.getRpcApiHttpHost();
@@ -86,12 +88,19 @@ test(testCase, async (t: Test) => {
     // pre-provision keychain with mock backend holding the private key of the
     // test account that we'll reference while sending requests with the
     // signing credential pointing to this keychain entry.
-    backend: new Map([[TestTokenJSON.contractName, TestTokenJSON]]),
+    backend: new Map([
+      [TestTokenJSON.contractName, JSON.stringify(TestTokenJSON)],
+    ]),
     logLevel,
   });
-  keychainPlugin.set(DemoHelperJSON.contractName, DemoHelperJSON);
-  keychainPlugin.set(HashTimeLockJSON.contractName, HashTimeLockJSON);
-
+  keychainPlugin.set(
+    DemoHelperJSON.contractName,
+    JSON.stringify(DemoHelperJSON),
+  );
+  keychainPlugin.set(
+    HashTimeLockJSON.contractName,
+    JSON.stringify(HashTimeLockJSON),
+  );
   const factory = new PluginFactoryLedgerConnector({
     pluginImportType: PluginImportType.Local,
   });
@@ -265,7 +274,7 @@ test(testCase, async (t: Test) => {
     web3SigningCredential,
     gas: estimatedGas,
   };
-  const res = await api.newContract(request);
+  const res = await api.newContractV1(request);
   t.equal(res.status, 200, "response status is 200 OK");
 
   t.comment("Get account balance");
@@ -306,7 +315,7 @@ test(testCase, async (t: Test) => {
     connectorId,
     keychainId,
   };
-  const resRefund = await api.refund(refundRequest);
+  const resRefund = await api.refundV1(refundRequest);
   t.equal(resRefund.status, 200, "response status is 200 OK");
 
   t.comment("Get account balance");
@@ -324,7 +333,7 @@ test(testCase, async (t: Test) => {
     "balance of account is 100 OK",
   );
   t.comment("Get status of HTLC");
-  const resStatus = await api.getSingleStatus(
+  const resStatus = await api.getSingleStatusV1(
     id,
     web3SigningCredential,
     connectorId,
@@ -357,12 +366,19 @@ test("Test invalid refund with invalid time", async (t: Test) => {
     // pre-provision keychain with mock backend holding the private key of the
     // test account that we'll reference while sending requests with the
     // signing credential pointing to this keychain entry.
-    backend: new Map([[TestTokenJSON.contractName, TestTokenJSON]]),
+    backend: new Map([
+      [TestTokenJSON.contractName, JSON.stringify(TestTokenJSON)],
+    ]),
     logLevel,
   });
-  keychainPlugin.set(DemoHelperJSON.contractName, DemoHelperJSON);
-  keychainPlugin.set(HashTimeLockJSON.contractName, HashTimeLockJSON);
-
+  keychainPlugin.set(
+    DemoHelperJSON.contractName,
+    JSON.stringify(DemoHelperJSON),
+  );
+  keychainPlugin.set(
+    HashTimeLockJSON.contractName,
+    JSON.stringify(HashTimeLockJSON),
+  );
   const factory = new PluginFactoryLedgerConnector({
     pluginImportType: PluginImportType.Local,
   });
@@ -536,7 +552,7 @@ test("Test invalid refund with invalid time", async (t: Test) => {
     web3SigningCredential,
     gas: estimatedGas,
   };
-  const res = await api.newContract(request);
+  const res = await api.newContractV1(request);
   t.equal(res.status, 200, "response status is 200 OK");
 
   t.comment("Get HTLC Id from DemoHelper");
@@ -566,7 +582,7 @@ test("Test invalid refund with invalid time", async (t: Test) => {
       connectorId,
       keychainId,
     };
-    const resRefund = await api.refund(refundRequest);
+    const resRefund = await api.refundV1(refundRequest);
     t.equal(resRefund.status, 400, "response status is 400");
   } catch (error) {
     t.equal(error.response.status, 400, "response status is 400");
@@ -582,11 +598,5 @@ test("Test invalid refund with invalid time", async (t: Test) => {
     params: [firstHighNetWorthAccount],
   });
   t.equal(responseFinalBalance.callOutput, "90", "balance of account is 90 OK");
-  t.end();
-});
-
-test("AFTER " + testCase, async (t: Test) => {
-  const pruning = pruneDockerAllIfGithubAction({ logLevel });
-  await t.doesNotReject(pruning, "Pruning did not throw OK");
   t.end();
 });