+++ webpack.prod.node.js
@@ -25,7 +25,6 @@ module.exports = {
         test: /\.ts$/,
         exclude: /node_modules/,
         use: [
-          "cache-loader",
           {
             loader: "ts-loader",
             options: {
@@ -39,7 +38,6 @@ module.exports = {
         test: /\.(js|ts)$/,
         enforce: "pre",
         use: [
-          "cache-loader",
           {
             loader: "source-map-loader",
           },
@@ -74,5 +72,6 @@ module.exports = {
     npm: "npm",
     "fabric-client": "fabric-client",
     "fabric-ca-client": "fabric-ca-client",
+    "@azure/identity": "@azure/identity",
   },
 };